feat: Added functions to calculate critical points
In this commit, I added two new functions to the
`critical_points_2d.py` module:

1. `find_critical_points`: This function takes a
2D function as input and returns a list of
critical points, i.e., points where the gradient
of the function is zero or undefined.
2. `plot_critical_points`: This function takes a
2D function and a list of critical points as input
and plots the function along with the critical points.

These functions use the `sympy` library to calculate
the gradient of the input function and find
its critical points. The `matplotlib` library
is used to visualize the function and the critical
points.